requirements.txt
streamlit
pymongo
python-dotenv
langchain
plotly
pandas
langchain-community
youtube-search-python
requests
gTTS